fix a peculiar cuda/thrust related test error
tl_tensor_cuda.cu have lots of instanitiations of thrust::sort_by_key(),
and it turns out that the in first start of the test program, the time spent
on blocking on cudaMalloc() has a strong positive correlation with the number of
instanitiations of thrust::sort_by_key(). The correlation with other thrust functions
has not been tested yet.

TODO: better to replace the thrust functions with our hand-written functions.